renaming context file to index
