+++ src/pages/login/hooks/__tests__/useLoginPage.test.js
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 import * as useCustomHook from '../useLoginSchema';
 import { useLoginPage } from '../useLoginPage';
-import * as useUserState from '../../../../global/redux/user/context';
+import * as useUserState from '../../../../global/redux/user';
 import { USER_ACTION_TYPES } from '../../../../global/redux/user/interface';
 
 const schemaMock = {
