+++ src/pages/login/hooks/useLoginPage.tsx
@@ -4,7 +4,7 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { useLoginRequest } from '../../../services/api/login/useLoginRequest';
 import { LoginInput } from '../../../services/api/login/interface';
 import { IFormInputNames } from '../interfaces/FormInputNames';
-import { useUserState } from '../../../global/redux/user/context';
+import { useUserState } from '../../../global/redux/user';
 import { USER_ACTION_TYPES } from '../../../global/redux/user/interface';
 
 export const useLoginPage = () => {
