+++ src/services/api/login/__tests__/useLoginRequest.test.js
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { useLoginRequest } from '../useLoginRequest';
-import * as useUserState from '../../../../global/redux/user/context';
+import * as useUserState from '../../../../global/redux/user';
 import * as usePagesNavigation from '../../../../hooks/usePagesNavigation';
 import { USER_ACTION_TYPES } from '../../../../global/redux/user/interface';
